git issues

+++ src/main/java/uk/ac/cf/cs/ons/skillsdb/skillsdb/users/UserController.java
@@ -83,10 +83,19 @@ public class UserController {
 
     @RequestMapping("deleteUser")
     public String deleteUser(@RequestParam("id") long id){
-        service.deleteUserById(id);
-        asService.deleteAssociatedSkillsByUserId(id);
-        return "index";
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User requesteduser = service.findUserByIndex(id).get();
+        String loggedinuser = authentication.getName();
+
+        if (requesteduser != null && requesteduser.getUsername().equals(loggedinuser)) {
+            service.deleteUserById(id);
+            asService.deleteAssociatedSkillsByUserId(id);
+
+            return "index";
+        } else {
+            return "404";
+        }
 
     }
 
@@ -108,11 +117,10 @@ public class UserController {
 
 
 
-
-        modelAndView.setViewName("users/profile");
-        model.addAttribute("user", user);
-        modelAndView.addObject("username", user.getUsername());
-        modelAndView.addObject("uid", user.getId());
+            modelAndView.setViewName("users/profile");
+            model.addAttribute("user", user);
+            modelAndView.addObject("username", user.getUsername());
+            modelAndView.addObject("uid", user.getId());
